third_party/WebKit/LayoutTests/external/wpt/webrtc/RTCPeerConnection-addTransceiver-expected.txt

This is a testharness.js-based test.
FAIL addTransceiver() with string argument as invalid kind should throw TypeError assert_idl_attribute: property "addTransceiver" not found in prototype chain
FAIL addTransceiver('audio') should return an audio transceiver assert_idl_attribute: property "addTransceiver" not found in prototype chain
FAIL addTransceiver('video') should return a video transceiver assert_idl_attribute: property "addTransceiver" not found in prototype chain
FAIL addTransceiver() with direction sendonly should have result transceiver.direction be the same pc.addTransceiver is not a function
FAIL addTransceiver() with direction inactive should have result transceiver.direction be the same pc.addTransceiver is not a function
FAIL addTransceiver() with invalid direction should throw TypeError assert_idl_attribute: property "addTransceiver" not found in prototype chain
FAIL addTransceiver(track) should have result with sender.track be given track assert_idl_attribute: Expect pc to have addTransceiver() method property "addTransceiver" not found in prototype chain
FAIL addTransceiver(track) multiple times should create multiple transceivers assert_idl_attribute: Expect pc to have addTransceiver() method property "addTransceiver" not found in prototype chain
FAIL addTransceiver() with rid containing invalid non-alphanumeric characters should throw TypeError assert_idl_attribute: property "addTransceiver" not found in prototype chain
FAIL addTransceiver() with rid longer than 16 characters should throw TypeError assert_idl_attribute: property "addTransceiver" not found in prototype chain
FAIL addTransceiver() with valid rid value should succeed pc.addTransceiver is not a function
FAIL addTransceiver() with readonly ssrc set should throw InvalidAccessError assert_throws: function "() =>
      pc.addTransceiver('audio', {
        sendEncodings: [{
          ssrc: 2
        }]
      })" threw object "TypeError: pc.addTransceiver is not a function" that is not a DOMException InvalidAccessError: property "code" is equal to undefined, expected 15
FAIL addTransceiver() with readonly rtx set should throw InvalidAccessError assert_throws: function "() =>
      pc.addTransceiver('audio', {
        sendEncodings: [{
          rtx: {
            ssrc: 2
          }
        }]
      })" threw object "TypeError: pc.addTransceiver is not a function" that is not a DOMException InvalidAccessError: property "code" is equal to undefined, expected 15
FAIL addTransceiver() with readonly fec set should throw InvalidAccessError assert_throws: function "() =>
      pc.addTransceiver('audio', {
        sendEncodings: [{
          fec: {
            ssrc: 2
          }
        }]
      })" threw object "TypeError: pc.addTransceiver is not a function" that is not a DOMException InvalidAccessError: property "code" is equal to undefined, expected 15
FAIL addTransceiver() with valid sendEncodings should succeed pc.addTransceiver is not a function
Harness: the test ran to completion.
